### 2026-01-16 ba15d62
JOSS: clarify AI usage for paper editing.
