### 2026-01-17 #9553 @boisgera
Hi @whoenig,

One quick question to help me find adequate reviewers for your submission: do you think that a reviewer with a UAV background and ROS2 knowledge can properly assess your work or should I restrict my search to potential reviewers who own Bitcraze robots? (that may be a bit more tricky).

Best regards,

Sébastien
