### 2026-01-18 #9553 @whoenig
I think UAV background and ROS2 knowledge suffices. Some of the suggested reviewers from https://github.com/openjournals/joss-reviews/issues/9553#issuecomment-3639058065 have experience in all three, but I don't think this is necessary for a review.
Thanks,
Wolfgang
