### 2026-01-28 50e7824
add explicit "state of field" section
