### 2026-01-28 IMRCLab/crazyswarm2#817 @llanesc
Consistency in backend cpp and cflib
I was exploring the use of the cpp backend for CrazySim since I read that the cpp backend has better latency and external pose management for hardware. However, I noticed a few issues with consistency between cflib and cpp servers.

1. For the odometry default topic the velocity estimate from the kalman.statePX is in the body frame while the stateEstimateZ.vx is in the global frame. The former is being used by cflib and the latter is used in the cpp server. The pose orientation is also pulling from different log variable groups as well.
2. Additionally, the python server is publishing TransformStamped in the odom callback while the cpp server is not.
3. Is there a way to defined custom topics with compressed log variables? I think currently it's using the LogDataGeneric which assumes the data is in floats. Perhaps a check if a capital Z is at the end of the log variable name to uncompress the variable? I tried to use the compressed acceleration log variables, but it wasn't publishing at all. I had to manually add it as a default. Perhaps the defaults can be extended to pose, pose_compressed, odom, odom_compressed?
4. I think it's quite useful to have the ability to perform offboard attitude setpoint control for some research applications. I remember we discussed this in the past and the conclusion was that the attitude setpoint was dependent on a seperate mode parameter to switch between absolute angle or angular rate. However, _cmd_vel_legacy_changed kind of already assumes this as well. Perhaps we can explore a temporary solution for users to have attitude setpoint commands.
