+++ crazyflie/src/crazyflie_server.cpp
@@ -181,6 +181,7 @@ public:
 
 
     service_emergency_ = node->create_service<Empty>(name + "/emergency", std::bind(&CrazyflieROS::emergency, this, _1, _2),  get_service_qos(), callback_group_cf_srv);
+    service_reboot_ = node->create_service<Empty>(name + "/reboot", std::bind(&CrazyflieROS::reboot, this, _1, _2), get_service_qos(), callback_group_cf_srv);
     service_start_trajectory_ = node->create_service<StartTrajectory>(name + "/start_trajectory", std::bind(&CrazyflieROS::start_trajectory, this, _1, _2), get_service_qos(), callback_group_cf_srv);
     service_takeoff_ = node->create_service<Takeoff>(name + "/takeoff", std::bind(&CrazyflieROS::takeoff, this, _1, _2), get_service_qos(), callback_group_cf_srv);
     service_land_ = node->create_service<Land>(name + "/land", std::bind(&CrazyflieROS::land, this, _1, _2), get_service_qos(), callback_group_cf_srv);
@@ -188,7 +189,7 @@ public:
     service_upload_trajectory_ = node->create_service<UploadTrajectory>(name + "/upload_trajectory", std::bind(&CrazyflieROS::upload_trajectory, this, _1, _2), get_service_qos(), callback_group_cf_srv);
     service_notify_setpoints_stop_ = node->create_service<NotifySetpointsStop>(name + "/notify_setpoints_stop", std::bind(&CrazyflieROS::notify_setpoints_stop, this, _1, _2), get_service_qos(), callback_group_cf_srv);
     service_arm_ = node->create_service<Arm>(name + "/arm", std::bind(&CrazyflieROS::arm, this, _1, _2), get_service_qos(), callback_group_cf_srv);
-
+    service_reboot_ = node->create_service<Empty>(name + "/reboot", std::bind(&CrazyflieROS::reboot, this, _1, _2), service_qos, callback_group_cf_srv);
     // Topics
 
     subscription_cmd_vel_legacy_ = node->create_subscription<geometry_msgs::msg::Twist>(name + "/cmd_vel_legacy", rclcpp::SystemDefaultsQoS(), std::bind(&CrazyflieROS::cmd_vel_legacy_changed, this, _1), sub_opt_cf_cmd);
@@ -716,6 +717,13 @@ private:
     cf_.emergencyStop();
   }
 
+  void reboot(const std::shared_ptr<Empty::Request> request,
+            std::shared_ptr<Empty::Response> response)
+  {
+    RCLCPP_INFO(logger_, "[%s] reboot()", name_.c_str());
+    cf_.reboot();
+  }
+
   void start_trajectory(const std::shared_ptr<StartTrajectory::Request> request,
                         std::shared_ptr<StartTrajectory::Response> response)
   {
@@ -1051,6 +1059,7 @@ private:
   tf2_ros::TransformBroadcaster tf_broadcaster_;
 
   rclcpp::Service<Empty>::SharedPtr service_emergency_;
+  rclcpp::Service<Empty>::SharedPtr service_reboot_;
   rclcpp::Service<StartTrajectory>::SharedPtr service_start_trajectory_;
   rclcpp::Service<Takeoff>::SharedPtr service_takeoff_;
   rclcpp::Service<Land>::SharedPtr service_land_;
@@ -1221,16 +1230,26 @@ public:
         }
       }
     }
-
+    bool relative_pose = false;
+    auto it = parameter_overrides.find("relative");
+    if (it != parameter_overrides.end()) {
+      relative_pose = it->second.get<bool>();
+    }
     this->declare_parameter("poses_qos_deadline", 100.0f);
[…]
