+++ crazyflie/CMakeLists.txt
@@ -127,6 +127,7 @@ install(PROGRAMS
   scripts/aideck_streamer.py
   scripts/gui.py
   scripts/flash.py
+  scripts/watch_dog.py
   DESTINATION lib/${PROJECT_NAME}
 )
 

[…]
