### 2026-01-29 IMRCLab/crazyswarm2#817 @llanesc
3. Oh okay I thought compressed variables couldn't go through the LogGeneric because I have a bug on CrazySim where sometimes some variables don't get published, but it worked when I did it through the default hard coding. I think there's some lost packets and the log group doesn't get requested from the firmware. I just checked and it does work.
4. That looks great! It would be nice to have this manual mode. Although perhaps we can request the CRTP_PROTOCOL_VERSION in the server and only allow this feature for specific CRTP versions and notify the user to update the firmware to have CRTP_PROTOCOL_VERSION >= 10 since it was added somewhere between 9 and 10. Maybe we can call it manualSetpoint to be consistent with the firmware although I agree the name is a bit strange. It would be nice if they could have a MAJOR and MINOR crtp version so that small changes like this can be differentiated when checking if a feature exists in a user's version with external software that uses CRTP.
