### 2026-02-10 #9947 @editorialbot
[REVIEW]: Crazyswarm2: A ROS 2-based Stack for Bitcraze Crazyflie Multirotor Robots
**Submitting author:** @whoenig (<a href="http://orcid.org/0000-0002-0773-028X">Wolfgang Hoenig</a>)
**Repository:** https://github.com/IMRCLab/crazyswarm2
**Branch with paper.md** (empty if default branch): joss
**Version:** 1.0.3
