### 2026-02-12 IMRCLab/crazyswarm2#817 @whoenig
Re 2: We should add it to the cpp-backend.
Re 3: Seems to be solved after the discussion
Re 4: This is a feature request.
