### 2026-02-16 #9947 @boisgera
For the record: @mrpollo disclosed privately with JOSS the details of a potential conflict of interest. Given the specific context, and after consultation of the JOSS editors (including the track editor in chief), we have decided to waive this COI.
