+++ crazyflie_examples/crazyflie_examples/cn_cbf.py
@@ -20,60 +20,6 @@ from torch.nn.utils import vector_to_parameters
 from scipy import ndimage
 from ament_index_python.packages import get_package_share_directory
 
-BASELINE = True
-
-
-def executeTrajectory(
-    timeHelper, cf, trajpath, rate=100, offset=np.zeros(3), yawrate=0.0
-):
-    traj = Trajectory()
-    traj.loadcsv(trajpath)
-
-    start_time = timeHelper.time()
-    yaw = 0
-    while not timeHelper.isShutdown():
-        t = timeHelper.time() - start_time
-        if t > traj.duration:
-            break
-
-        e = traj.eval(t)
-        yaw = yaw + yawrate * 1 / rate
-        omega = e.omega
-        omega[2] = yawrate
-        cf.cmdFullState(
-            e.pos + np.array(cf.initialPosition) + offset, e.vel, e.acc, yaw, omega
-        )
-        print(yaw)
-
-        timeHelper.sleepForRate(rate)
-
-
-# def executeTrajectory2(timeHelper, cf, rate=100, offset=np.zeros(3), yawrate=0.0):
-
-#     start_time = timeHelper.time()
-#     yaw = 0
-#     pos = np.zeros(3)
-#     vel = np.zeros(3)
-#     acc = np.array([0])
-#     while not timeHelper.isShutdown():
-#         t = timeHelper.time() - start_time
-#         if t > 5.0:
-#             break
-
-#         e = traj.eval(t)
-#         yaw = yaw + yawrate * 1/rate
-#         omega = e.omega
-#         omega[2] = yawrate
-#         cf.cmdFullState(
-#             e.pos + np.array(cf.initialPosition) + offset,
-#             e.vel,
-#             e.acc,
-#             yaw,
-#             omega)
-#         print(yaw)
-
-#         timeHelper.sleepForRate(rate)
-
 
 class Sin(nn.Module):
     """Sine activation function"""
@@ -103,47 +49,7 @@ def get_func(func_name: str):
     return func_bank[func_name]
 
 
-class Hypernetwork(nn.Module):
-    """Predicts parameters of the main network based on a 2D input."""
-
-    def __init__(self, input_size: int, output_size: int) -> None:
-        super().__init__()
-        n = 8
-        self.backbone = nn.Sequential(
-            nn.Conv2d(input_size, n, (3, 3), (1, 1), "valid", bias=False),
-            nn.SELU(),
-            nn.Conv2d(n, n, (3, 3), (1, 1), "valid", bias=False),
-            nn.SELU(),
-            nn.AvgPool2d((2, 2), (1, 1)),
-            nn.Conv2d(n, 2 * n, (3, 3), (2, 2), "valid", bias=False),
-            # (2*n, 47, 47)
-            nn.SELU(),
-            nn.Conv2d(2 * n, 2 * n, (3, 3), (1, 1), "valid", bias=False),
-            nn.SELU(),
-            nn.AvgPool2d((2, 2), (1, 1)),
-            nn.Conv2d(2 * n, 4 * n, (2, 2), (2, 2), "valid", bias=False),
-            # (4*n, 22, 22)
-            nn.SELU(),
-            nn.Conv2d(4 * n, 4 * n, (3, 3), (2, 2), "valid", bias=False),
-            nn.SELU(),
-            nn.Conv2d(4 * n, 8 * n, (2, 2), (2, 2), "valid", bias=False),
-            # (8*n, 4, 4)
-            nn.SELU(),
-            nn.Flatten(),
-        )
-
-        self.head = nn.Sequential(
-            nn.Linear(1024, 1024),
-            nn.SELU(),
-            nn.Linear(1024, output_size),
-        )
-
[…]
