### 2026-02-18 7683be2
CN-CBF initial implementation
