+++ crazyflie_examples/crazyflie_examples/cn_cbf.py
@@ -6,6 +6,7 @@ from crazyflie_py import Crazyswarm
 from crazyflie_py.uav_trajectory import Trajectory
 import numpy as np
 from nav_msgs.msg import OccupancyGrid
+from std_msgs.msg import Float32
 import rclpy
 from rclpy.node import Node
 
@@ -116,6 +117,7 @@ def softmin_torch(x, beta=10.0, dim=-1):
 
 class ORN_CBF:
     def __init__(self):
+        # Configuration parameters
         self.alpha = 4.0
         self.beta = 4.0
         self.r_min = 0.7
@@ -134,8 +136,10 @@ class ORN_CBF:
             ],
         }
 
+        # Define system dynamics
         self.system_dynamics = DoubleIntegrator2D()
 
+        # Define the MLP model and load weights
         pkg_share_dir = get_package_share_directory("crazyflie_examples")
         self.mlp = MLP(config=self.mlp_config).to(self.mlp_device)
         mlp_weights_path = os.path.join(
@@ -193,98 +197,115 @@ class ORN_CBF:
         self.qp_solver = ca.qpsol("qp_solver", "qpoases", qp, qp_opts)
 
         # Initialize variables
-        self.people_msg_old = []
-        self.people_msg_time_old = None
-        self.people_states = np.array([])
-        self.people_states_dot = np.array([])
+        self.obstacles_msg_old = []
+        self.obstacles_msg_time_old = None
+        self.obstacles_states = np.array([])
+        self.obstacles_states_dot = np.array([])
         self.robot_state = np.zeros(self.system_dynamics.nx)
         self.nom_control = np.zeros(self.system_dynamics.nu)
 
-    def update_main_net(self, costmap_msg: OccupancyGrid):
-        # Update the main neural network based on the latest costmap
-        costmap = np.reshape(
-            np.array(costmap_msg.data, dtype=np.bool_),
-            (costmap_msg.info.width, costmap_msg.info.height),
-        ).T
+    def obstacles_callback(self, msg):
+        obstacles_msg_time = self.get_clock().now().nanoseconds / 1e9
+        obstacles_states = []
+        obstacles_states_dot = []
+        for obstacle in msg.obstacles:
+            obstacle_state = np.array(
+                [
+                    obstacle.position.x,
+                    obstacle.position.y,
+                    obstacle.velocity.x,
+                    obstacle.velocity.y,
+                ]
+            )
+            obstacle_state_dot = np.zeros_like(obstacle_state)
+            for obstacle_old in self.obstacles_msg_old:
+                if obstacle.name == obstacle_old.name:
+                    obstacle_state_old = np.array(
+                        [
+                            obstacle_old.position.x,
+                            obstacle_old.position.y,
+                            obstacle_old.velocity.x,
+                            obstacle_old.velocity.y,
+                        ]
+                    )
+                    dt = obstacles_msg_time - self.obstacles_msg_time_old
+                    obstacle_state_dot = (obstacle_state - obstacle_state_old) / (
+                        dt + 1e-6
+                    )
 
-        # TODO: check if the costmap position is correct
-        x_offset = costmap_msg.info.height * costmap_msg.info.resolution / 2.0
-        y_offset = costmap_msg.info.width * costmap_msg.info.resolution / 2.0
-        self.costmap_pos = np.array(
-            [
-                costmap_msg.info.origin.position.x + x_offset,
-                costmap_msg.info.origin.position.y + y_offset,
-            ]
-        )
[…]
