### 2026-02-18 795a47a
Added obstacles states interface
