### 2026-02-18 d257539
Added new interfaces
