### 2026-02-18 57a2b83
Added new interfaces
