### 2026-02-18 700e48e
renamed Obstacles.msg to ObstacleArray.msg
