### 2026-02-19 e06043c
Code refactored
