+++ crazyflie_examples/crazyflie_examples/cn_cbf.py
@@ -1,24 +1,17 @@
 #!/usr/bin/env python
 
 from pathlib import Path
-
 from crazyflie_py import Crazyswarm
 from crazyflie_py.uav_trajectory import Trajectory
 import numpy as np
-from nav_msgs.msg import OccupancyGrid
+from crazyflie_interfaces.msg import ObstacleArray
 from std_msgs.msg import Float32
-import rclpy
-from rclpy.node import Node
-
-import time
 import os
 import casadi as ca
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
 from torch import Tensor
-from torch.nn.utils import vector_to_parameters
-from scipy import ndimage
 from ament_index_python.packages import get_package_share_directory
 
 
@@ -70,7 +63,7 @@ class MLP(nn.Module):
 
 
 class DoubleIntegrator2D:
-    """2D Double Integrator model of a robot."""
+    """2D Double Integrator model class."""
 
     def __init__(self):
         # State space
@@ -107,11 +100,13 @@ class DoubleIntegrator2D:
 
 
 def softmin_numpy(x, beta=10.0, axis=None):
+    """Numpy implementation of the softmin function."""
     out = -(1.0 / beta) * np.log(np.sum(np.exp(-beta * x), axis=axis, keepdims=True))
     return out.squeeze()
 
 
-def softmin_torch(x, beta=10.0, dim=-1):
+def softmin_torch(x, beta=10.0, dim=None):
+    """PyTorch implementation of the softmin function."""
     return -torch.logsumexp(-beta * x, dim=dim) / beta
 
 
@@ -201,50 +196,37 @@ class CN_CBF:
         self.qp_solver = ca.qpsol("qp_solver", "qpoases", qp, qp_opts)
 
         # Initialize variables
-        self.obstacles_msg_old = []
-        self.obstacles_msg_time_old = None
         self.obstacles_states = np.array([])
         self.obstacles_states_dot = np.array([])
         self.robot_state = np.zeros(self.system_dynamics.nx)
         self.nom_control = np.zeros(self.system_dynamics.nu)
 
     def obstacles_callback(self, msg):
-        obstacles_msg_time = self.get_clock().now().nanoseconds / 1e9
         obstacles_states = []
         obstacles_states_dot = []
         for obstacle in msg.obstacles:
             obstacle_state = np.array(
                 [
-                    obstacle.position.x,
-                    obstacle.position.y,
-                    obstacle.velocity.x,
-                    obstacle.velocity.y,
+                    obstacle.pose.position.x,
+                    obstacle.pose.position.y,
+                    obstacle.twist.linear.x,
+                    obstacle.twist.linear.y,
                 ]
             )
-            obstacle_state_dot = np.zeros_like(obstacle_state)
-            for obstacle_old in self.obstacles_msg_old:
-                if obstacle.name == obstacle_old.name:
-                    obstacle_state_old = np.array(
-                        [
-                            obstacle_old.position.x,
-                            obstacle_old.position.y,
-                            obstacle_old.velocity.x,
-                            obstacle_old.velocity.y,
-                        ]
-                    )
-                    dt = obstacles_msg_time - self.obstacles_msg_time_old
-                    obstacle_state_dot = (obstacle_state - obstacle_state_old) / (
-                        dt + 1e-6
-                    )
+            obstacle_state_dot = np.array(
+                [
[…]
