### 2026-02-19 f919c20
Added implementation of C3BF method
