+++ crazyflie_examples/crazyflie_examples/c3bf.py
@@ -1,0 +1,296 @@
+#!/usr/bin/env python
+
+from pathlib import Path
+from crazyflie_py import Crazyswarm
+from crazyflie_py.uav_trajectory import Trajectory
+import numpy as np
+from crazyflie_interfaces.msg import ObstacleArray
+import os
+import casadi as ca
+
+
+class DoubleIntegrator2D:
+    """2D Double Integrator model class."""
+
+    def __init__(self):
+        # State space
+        self.nx = 4
+        self.x_min = [-ca.inf, -ca.inf, -2.0, -2.0]
+        self.x_max = [ca.inf, ca.inf, 2.0, 2.0]
+        self.A = ca.DM(
+            [
+                [0.0, 0.0, 1.0, 0.0],
+                [0.0, 0.0, 0.0, 1.0],
+                [0.0, 0.0, 0.0, 0.0],
+                [0.0, 0.0, 0.0, 0.0],
+            ]
+        )
+
+        # Control space
+        self.nu = 2
+        self.u_min = [-2.0, -2.0]
+        self.u_max = [2.0, 2.0]
+        self.B = ca.DM(
+            [
+                [0.0, 0.0],
+                [0.0, 0.0],
+                [1.0, 0.0],
+                [0.0, 1.0],
+            ]
+        )
+
+    def open_loop_dynamics(self, x):
+        return self.A @ x
+
+    def control_jacobian(self, x):
+        return self.B
+
+
+class C3BF:
+    """Collision Cone Control Barrier Function (C3BF) class."""
+
+    def __init__(self):
+        # Configuration parameters
+        self.alpha = 3.0
+        self.max_num_obstacles = 5
+        self.obstacle_state_size = 4  # [x, y, vx, vy]
+        self.r_min = 0.7
+
+        # Define system dynamics
+        self.system_dynamics = DoubleIntegrator2D()
+
+        # Define CBF-QP and create QP solver
+        system_state = ca.MX.sym("system_state", self.system_dynamics.nx)
+        nom_control = ca.MX.sym("nom_control", self.system_dynamics.nu)
+        control = ca.MX.sym("control", self.system_dynamics.nu)
+        obstacle_states = ca.MX.sym(
+            "obstacle_states", self.max_num_obstacles, self.obstacle_state_size
+        )
+        obstacle_states_dot = ca.MX.sym(
+            "obstacle_states_dot", self.max_num_obstacles, self.obstacle_state_size
+        )
+        slack_var = ca.MX.sym("slack_var", self.max_num_obstacles)
+
+        # QP variables
+        qp_var = ca.vertcat(control, slack_var)
+
+        # QP objective
+        qp_obj = ca.dot(
+            (nom_control - control), (nom_control - control)
+        ) + 1e6 * ca.dot(slack_var, slack_var)
+
+        # QP constraints
+        f = self.system_dynamics.open_loop_dynamics(system_state)
+        g = self.system_dynamics.control_jacobian(system_state)
+
+        qp_constr = []
+        for i in range(self.max_num_obstacles):
+            p = obstacle_states[i, :2].T - system_state[:2]
+            v = obstacle_states[i, 2:].T - system_state[2:]
+            h = ca.dot(p, v) + ca.norm_2(v) * ca.sqrt(
+                ca.fmax(0.01, ca.norm_2(p) ** 2 - self.r_min**2)
+            )
+            grad_h_x = ca.gradient(h, system_state)
+            grad_h_o = ca.gradient(h, obstacle_states)[i, :].T
+            cbf_constr = (
+                grad_h_x.T @ (f + g @ control)
+                + grad_h_o.T @ obstacle_states_dot[i, :].T
[…]
