### 2026-02-19 8466fd5
bugfix in obstacles callback
