+++ crazyflie_examples/crazyflie_examples/dho_cbf.py
@@ -1,0 +1,293 @@
+#!/usr/bin/env python
+
+from pathlib import Path
+from crazyflie_py import Crazyswarm
+from crazyflie_py.uav_trajectory import Trajectory
+import numpy as np
+from crazyflie_interfaces.msg import ObstacleArray
+import casadi as ca
+
+
+class DoubleIntegrator2D:
+    """2D Double Integrator model class."""
+
+    def __init__(self):
+        # State space
+        self.nx = 4
+        self.x_min = [-ca.inf, -ca.inf, -2.0, -2.0]
+        self.x_max = [ca.inf, ca.inf, 2.0, 2.0]
+        self.A = ca.DM(
+            [
+                [0.0, 0.0, 1.0, 0.0],
+                [0.0, 0.0, 0.0, 1.0],
+                [0.0, 0.0, 0.0, 0.0],
+                [0.0, 0.0, 0.0, 0.0],
+            ]
+        )
+
+        # Control space
+        self.nu = 2
+        self.u_min = [-2.0, -2.0]
+        self.u_max = [2.0, 2.0]
+        self.B = ca.DM(
+            [
+                [0.0, 0.0],
+                [0.0, 0.0],
+                [1.0, 0.0],
+                [0.0, 1.0],
+            ]
+        )
+
+    def open_loop_dynamics(self, x):
+        return self.A @ x
+
+    def control_jacobian(self, x):
+        return self.B
+
+
+class DHO_CBF:
+    """Dynamic Higher-Order Control Barrier Function (DHO-CBF) class."""
+
+    def __init__(self):
+        # Configuration parameters
+        self.k1 = 3.0
+        self.k2 = 2.0
+        self.max_num_obstacles = 5
+        self.obstacle_state_size = 4  # [x, y, vx, vy]
+        self.r_min = 0.7
+
+        # Define system dynamics
+        self.system_dynamics = DoubleIntegrator2D()
+
+        # Define CBF-QP and create QP solver
+        system_state = ca.MX.sym("system_state", self.system_dynamics.nx)
+        nom_control = ca.MX.sym("nom_control", self.system_dynamics.nu)
+        control = ca.MX.sym("control", self.system_dynamics.nu)
+        obstacle_states = ca.MX.sym(
+            "obstacle_states", self.max_num_obstacles, self.obstacle_state_size
+        )
+        obstacle_states_dot = ca.MX.sym(
+            "obstacle_states_dot", self.max_num_obstacles, self.obstacle_state_size
+        )
+        slack_var = ca.MX.sym("slack_var", self.max_num_obstacles)
+
+        # QP variables
+        qp_var = ca.vertcat(control, slack_var)
+
+        # QP objective
+        qp_obj = ca.dot(
+            (nom_control - control), (nom_control - control)
+        ) + 1e6 * ca.dot(slack_var, slack_var)
+
+        # QP constraints
+        qp_constr = []
+        for i in range(self.max_num_obstacles):
+            p = obstacle_states[i, :2].T - system_state[:2]
+            v = obstacle_states[i, 2:].T - system_state[2:]
+            a = obstacle_states_dot[i, 2:].T - control
+
+            h = ca.norm_2(p) - self.r_min
+            h_dot = (p.T @ v) / ca.norm_2(p)
+            h_ddot = (p.T @ a + v.T @ v) / ca.norm_2(p) - (p.T @ v) ** 2 / ca.norm_2(
+                p
+            ) ** 3
+
+            cbf_constr = h_ddot + (self.k1 + self.k2) * h_dot + self.k1 * self.k2 * h
+            qp_constr.append(cbf_constr + slack_var[i])
+        qp_constr = ca.vertcat(*qp_constr)
+
+        # QP parameters
[…]
