### 2026-02-19 d630603
initial obstacle related code (untested)
