+++ crazyflie_examples/setup.cfg
@@ -11,3 +11,8 @@ console_scripts =
     infinite_flight = crazyflie_examples.infinite_flight:main
     arming = crazyflie_examples.arming:main
     orn_cbf = crazyflie_examples.orn_cbf:main
+    obstacle_publisher_from_cf = crazyflie_examples.obstacle_publisher_from_cf:main
+    obstacle_fly = crazyflie_examples.obstacle_fly:main
+    c3bf = crazyflie_examples.c3bf:main
+    cn_cbf = crazyflie_examples.cn_cbf:main
+    dho_cbf = crazyflie_examples.dho_cbf:main
