+++ crazyflie_examples/crazyflie_examples/obstacle_fly.py
@@ -11,7 +11,7 @@ def main():
     swarm = Crazyswarm()
     timeHelper = swarm.timeHelper
 
-    cfs = ["cf5"]
+    cfs = ["cf1"]
     TIMESCALE = 1.0
 
     allcfs = swarm.allcfs
@@ -23,23 +23,28 @@ def main():
         trajs.append(traj)
 
     for idx, cf in enumerate(cfs):
-        swarm.crazyfliesByName[cf].uploadTrajectory(0, 0, trajs[idx % len(trajs)])
+        allcfs.crazyfliesByName[cf].uploadTrajectory(0, 0, trajs[idx % len(trajs)])
 
     for cf in cfs:
-        swarm.crazyfliesByName[cf].takeoff(targetHeight=1.0, duration=2.0)
+        allcfs.crazyfliesByName[cf].takeoff(targetHeight=0.5, duration=2.0)
     timeHelper.sleep(3.0)
     
     for cf in cfs:
-        pos = np.array(swarm.crazyfliesByName[cf].initialPosition) + np.array([0.0, 0.0, 1.0])
-        swarm.crazyfliesByName[cf].goTo(pos, 0, 2.0)
+        pos = np.array(allcfs.crazyfliesByName[cf].initialPosition) + np.array([0.0, 0.0, 0.5])
+        allcfs.crazyfliesByName[cf].goTo(pos, 0, 2.0)
     timeHelper.sleep(2.5)
 
-    for cf in cfs:
-        swarm.crazyfliesByName[cf].startTrajectory(0, TIMESCALE)
-    timeHelper.sleep(max([t.duration for t in trajs]) * TIMESCALE + 2.0)
+    for i in range(5):
+        for cf in cfs:
+            allcfs.crazyfliesByName[cf].startTrajectory(0, TIMESCALE)
+        timeHelper.sleep(max([t.duration for t in trajs]) * TIMESCALE + 2.0)
+
+        for cf in cfs:
+            allcfs.crazyfliesByName[cf].startTrajectory(0, TIMESCALE, reverse=True)
+        timeHelper.sleep(max([t.duration for t in trajs]) * TIMESCALE + 2.0)
 
     for cf in cfs:
-        swarm.crazyfliesByName[cf].land(targetHeight=0.06, duration=2.0)
+        allcfs.crazyfliesByName[cf].land(targetHeight=0.06, duration=2.0)
     timeHelper.sleep(3.0)
 
 
