+++ crazyflie_examples/crazyflie_examples/dho_cbf.py
@@ -54,7 +54,7 @@ class DHO_CBF:
         self.k2 = 2.0
         self.max_num_obstacles = 5
         self.obstacle_state_size = 4  # [x, y, vx, vy]
-        self.r_min = 0.7
+        self.r_min = 0.3
 
         # Define system dynamics
         self.system_dynamics = DoubleIntegrator2D()
@@ -63,11 +63,11 @@ class DHO_CBF:
         system_state = ca.MX.sym("system_state", self.system_dynamics.nx)
         nom_control = ca.MX.sym("nom_control", self.system_dynamics.nu)
         control = ca.MX.sym("control", self.system_dynamics.nu)
-        obstacle_states = ca.MX.sym(
-            "obstacle_states", self.max_num_obstacles, self.obstacle_state_size
+        obstacles_states = ca.MX.sym(
+            "obstacles_states", self.max_num_obstacles, self.obstacle_state_size
         )
-        obstacle_states_dot = ca.MX.sym(
-            "obstacle_states_dot", self.max_num_obstacles, self.obstacle_state_size
+        obstacles_states_dot = ca.MX.sym(
+            "obstacles_states_dot", self.max_num_obstacles, self.obstacle_state_size
         )
         slack_var = ca.MX.sym("slack_var", self.max_num_obstacles)
 
@@ -82,9 +82,9 @@ class DHO_CBF:
         # QP constraints
         qp_constr = []
         for i in range(self.max_num_obstacles):
-            p = obstacle_states[i, :2].T - system_state[:2]
-            v = obstacle_states[i, 2:].T - system_state[2:]
-            a = obstacle_states_dot[i, 2:].T - control
+            p = obstacles_states[i, :2].T - system_state[:2]
+            v = obstacles_states[i, 2:].T - system_state[2:]
+            a = obstacles_states_dot[i, 2:].T - control
 
             h = ca.norm_2(p) - self.r_min
             h_dot = (p.T @ v) / ca.norm_2(p)
@@ -100,8 +100,8 @@ class DHO_CBF:
         qp_params = ca.vertcat(
             system_state,
             nom_control,
-            ca.vec(obstacle_states),
-            ca.vec(obstacle_states_dot),
+            ca.vec(obstacles_states),
+            ca.vec(obstacles_states_dot),
         )
 
         # QP options
@@ -134,11 +134,13 @@ class DHO_CBF:
         )
 
     def obstacles_callback(self, msg):
-        obstacles_states = []
-        obstacles_states_dot = []
+        self.obstacles_states = self.obstacles_states_init.copy()
+        self.obstacles_states_dot = np.zeros(
+            (self.max_num_obstacles, self.obstacle_state_size)
+        )
         for i in range(min(len(msg.obstacles), self.max_num_obstacles)):
             obstacle = msg.obstacles[i]
-            obstacle_state = np.array(
+            self.obstacles_states[i, :] = np.array(
                 [
                     obstacle.pose.position.x,
                     obstacle.pose.position.y,
@@ -146,7 +148,7 @@ class DHO_CBF:
                     obstacle.twist.linear.y,
                 ]
             )
-            obstacle_state_dot = np.array(
+            self.obstacles_states_dot[i, :] = np.array(
                 [
                     obstacle.twist.linear.x,
                     obstacle.twist.linear.y,
@@ -154,12 +156,6 @@ class DHO_CBF:
                     obstacle.accel.y,
                 ]
             )
-
[…]

[…]
