+++ crazyflie/config/motion_capture.yaml
@@ -20,7 +20,7 @@
           p2: [-0.042, -0.042, 0.0  ]  # back right  (M2)
           p3: [ 0.042, -0.042, 0.0  ]  # front right (M1)
       C10X:
-        offset: [0.0, 0.0, -0.1]
+        offset: [0.0, 0.0, 0.0]
         points:
           p0: [-0.0677304,0.0457213,-0.000254381]
           p1: [-0.0665792,-0.018327,-0.000410748]
