+++ crazyflie/config/crazyflies.yaml
@@ -31,31 +31,37 @@ robots:
     uri: radio://0/80/2M/E7E7E7E715
     initial_position: [0.0, 0.0, 0.0]
     type: C10X # see robot_types
+    role: None
   C22:
     enabled: false
     uri: radio://0/80/2M/E7E7E7E716
     initial_position: [0.0, 0.0, 0.0]
     type: C10X # see robot_types
+    role: None
   C23:
-    enabled: false
+    enabled: true
     uri: radio://0/80/2M/E7E7E7E717
     initial_position: [0.0, 0.0, 0.0]
     type: C10X # see robot_types
+    role: evader
   C24:
-    enabled: false
+    enabled: true
     uri: radio://0/80/2M/E7E7E7E718
     initial_position: [0.0, 0.0, 0.0]
     type: C10X # see robot_types
+    role: pursuer
   C25:
-    enabled: false
+    enabled: true
     uri: radio://0/80/2M/E7E7E7E719
     initial_position: [0.0, 0.0, 0.0]
     type: C10X # s
+    role: pursuer
   C26:
     enabled: true
     uri: radio://0/80/2M/E7E7E7E71A
     initial_position: [0.0, 0.0, 0.0]
     type: C10X # s
+    role: pursuer
 #####################################################################################
   C101:
     enabled: false
@@ -221,16 +227,30 @@ all:
     #   vyKi: 14.0
     #   vyKd: 1.0
     #   vyKFF: 0.0
-    posCtlPid:
-      xKp: 0.8
-    #   xKi: 0.15
-      xKd: 0.05
-      yKp: 0.8
-    #   yKi: 0.15
-      yKd: 0.05
+    # posCtlPid:
+    #   xKp: 1.0
+    #   xKi: 0.05
+    #   xKd: 0.1
+
+    #   yKp: 1.0
+    #   yKi: 0.05
+    #   yKd: 0.1
+
     #   zKp: 2.0
     #   zKi: 0.70
     #   zKd: 0.1
+
+    # --------- OLD
+      # xKp: 0.8
+      # xKi: 0.15
+      # xKd: 0.05
+      # yKp: 0.8
+      # yKi: 0.15
+      # yKd: 0.05
+      # zKp: 2.0
+      # zKi: 0.70
+      # zKd: 0.1
+    # -------------
     # ctrlMel:
     #   mass: 0.040 #mass for crazyflie with box
       # kp_xy: 0.1
@@ -262,4 +282,4 @@ all:
   broadcasts:
     num_repeats: 15 # number of times broadcast commands are repeated
     delay_between_repeats_ms: 1 # delay in milliseconds between individual repeats
-relative: true
+relative: false
