### 2026-02-25 c47415b
initial simple 2 UAV demo
