+++ crazyflie/launch/launch.py
@@ -126,10 +126,10 @@ def generate_launch_description():
                               default_value=default_rviz_config_path),
         DeclareLaunchArgument('backend', default_value='cpp'),
         DeclareLaunchArgument('debug', default_value='False'),
-        DeclareLaunchArgument('rviz', default_value='False'),
-        DeclareLaunchArgument('gui', default_value='True'),
+        DeclareLaunchArgument('rviz', default_value='True'),
+        DeclareLaunchArgument('gui', default_value='False'),
         DeclareLaunchArgument('teleop', default_value='True'),
-        DeclareLaunchArgument('mocap', default_value='True'),
+        DeclareLaunchArgument('mocap', default_value='False'),
         DeclareLaunchArgument('teleop_yaml_file', default_value=''),
         OpaqueFunction(function=parse_yaml),
         Node(
