+++ crazyflie_examples/crazyflie_examples/nice_hover.py
@@ -17,8 +17,11 @@ def main():
         pos = np.array(cf.initialPosition) + np.array([0, 0, Z])
         cf.goTo(pos, 0, 1.0)
 
-    print('press button to continue...')
-    swarm.input.waitUntilButtonPressed()
+    # print('press button to continue...')
+    # swarm.input.waitUntilButtonPressed()
+    allcfs.setParam('ctrlLee.indi', 1)
+    timeHelper.sleep(3.0)
+    allcfs.setParam('ctrlLee.indi', 0)
 
     allcfs.land(targetHeight=0.02, duration=1.0+Z)
     timeHelper.sleep(1.0+Z)
