### 2026-02-28 ab9af8c
Added cf8 to obstacle list
