### 2026-03-05 f26d907
autotuning kappa_f (untested)
