### 2026-03-06 c6ead90
updates for flow-only experiments
