+++ crazyflie/config/crazyflies.yaml
@@ -7,9 +7,23 @@ robots:
     uri: radio://0/80/2M/E7E7E7AA01
     initial_position: [0.0, 0.2, 0.0]
     type: cf21  # see robot_types
+    firmware_params:
+      kalman:
+        initialY: 0.2
+        resetEstimation: 1
+
+  aa02:
+    enabled: true
+    uri: radio://0/80/2M/E7E7E7AA02
+    initial_position: [0.0, -0.2, 0.0]
+    type: cf21  # see robot_types
+    firmware_params:
+      kalman:
+        initialY: -0.2
+        resetEstimation: 1
 
   bb01:
-    enabled: true
+    enabled: false
     uri: radio://0/80/2M/E7E7E7BB01
     initial_position: [0.0, -0.2, 0.0]
     type: cf21bl  # see robot_types
@@ -73,9 +87,9 @@ all:
     #   solidGreen: 0 # if set to solid color
     #   solidRed: 0 # if set to solid color
     #   headlightEnable: 0
-    locSrv:
-      extPosStdDev: 1e-3
-      extQuatStdDev: 0.5e-1
+    # locSrv:
+    #   extPosStdDev: 1e-3
+    #   extQuatStdDev: 0.5e-1
     # kalman:
     #   resetEstimation: 1
   # reference frame for on-board state-estimate of all drones (use robot_types/type_name to set per type, or
