+++ crazyflie_examples/setup.cfg
@@ -8,6 +8,7 @@ console_scripts =
     multi_trajectory = crazyflie_examples.multi_trajectory:main
     cmd_full_state = crazyflie_examples.cmd_full_state:main
     set_param = crazyflie_examples.set_param:main
-    swap = crazyflie_examples.swap:main
+    swap_basic = crazyflie_examples.swap_basic:main
+    swap_indi = crazyflie_examples.swap_indi:main
     infinite_flight = crazyflie_examples.infinite_flight:main
     arming = crazyflie_examples.arming:main
