### 2026-03-12 #9947 @boisgera
Hi @llanesc,

- wrt **Data sharing** and **Reproducibility** items in your checklist, I guess (?) that the paper contains no original data and no new result? If you can confirm that, we should check these items (according to the JOSS convention).

- I see that you still have [an open issue](https://github.com/IMRCLab/crazyswarm2/issues/817) in crazyswarm2 but afaict the discussion has addressed some of your points and triaged the others. Am I right to assume that the items that remain are non-blockers for the publication of the paper in JOSS?

Best,

Sébastien
