### 2026-05-26 f0fdbdd
CI: switch rolling from noble -> resolute
