+++ .github/workflows/ci-ros2.yml
@@ -32,7 +32,7 @@ jobs:
             ros_version: 2
 
           # Rolling Ridley  (June 2020 - Present)
-          - docker_image: rostooling/setup-ros-docker:ubuntu-noble-ros-rolling-ros-base-latest
+          - docker_image: rostooling/setup-ros-docker:ubuntu-resolute-ros-rolling-ros-base-latest
             ros_distribution: rolling
             ros_version: 2
 
@@ -52,6 +52,7 @@ jobs:
           # a python package globally by disabling the externally managed flag that is default on Ubuntu 24.04
         run: |
           sudo rm -f /usr/lib/python3.12/EXTERNALLY-MANAGED
+          sudo rm -f /usr/lib/python3.14/EXTERNALLY-MANAGED
           pip install rowan
 
       - uses: actions/checkout@v2
