### 2026-05-26 9134882
merge multiple target_link_libraries calls
