### 2026-05-26 04243b0
add lyrical to the CI
