+++ .github/workflows/ci-ros2.yml
@@ -17,6 +17,7 @@ jobs:
         ros_distribution:
           - humble
           - jazzy
+          - lyrical
           - rolling
 
         # Define the Docker image(s) associated with each ROS distribution.
@@ -31,6 +32,11 @@ jobs:
             ros_distribution: jazzy
             ros_version: 2
 
+          # Lyrical Luth (May 2026 - May 2031)
+          - docker_image: rostooling/setup-ros-docker:ubuntu-resolute-ros-lyrical-ros-base-latest
+            ros_distribution: lyrical
+            ros_version: 2
+          
           # Rolling Ridley  (June 2020 - Present)
           - docker_image: rostooling/setup-ros-docker:ubuntu-resolute-ros-rolling-ros-base-latest
             ros_distribution: rolling
