### 2026-05-26 c3eedaa
CI: manual dispatch
