+++ .github/workflows/ci-ros2.yml
@@ -5,6 +5,8 @@ on:
     branches: [ main ]
   pull_request:
     branches: [ main ]
+  # enables manual dispatch on github
+  workflow_dispatch:
 
 # Based on example provided at https://github.com/ros-tooling/setup-ros
 
@@ -63,7 +65,7 @@ jobs:
 
       - uses: actions/checkout@v2
       - name: build and test ROS 2
-        uses: ros-tooling/action-ros-ci@v0.3
+        uses: ros-tooling/action-ros-ci@v0.4
         with:
           package-name: |
             crazyflie
