+++ crazyflie_server_rs/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 
 [dependencies]
 tokio = { version = "1", features = ["macros", "rt", "rt-multi-thread", "time"] }
-rclrs = "0.6"
+rclrs = "0.7"
 crazyflie-link = "0.3.0"
 crazyflie-lib = "0.3.0"
 crazyflie_interfaces = "*"
