### 2026-05-26 5aba6c0
get working almost feature complete rust library
