+++ crazyflie_server_rs/src/main.rs
@@ -1,62 +1,495 @@
 use crazyflie_lib::Crazyflie;
-use rclrs::{Node, Service, Context, CreateBasicExecutor, SpinOptions, ServiceInfo};
-use crazyflie_interfaces::srv::{Takeoff, Takeoff_Request, Takeoff_Response};
+use rclrs::{Node, Service, Publisher, Subscription, Context, CreateBasicExecutor, SpinOptions};
+use crazyflie_interfaces::{
+    srv::{
+        Takeoff, Takeoff_Request, Takeoff_Response,
+        Land, Land_Request, Land_Response,
+        GoTo, GoTo_Request, GoTo_Response,
+        Arm, Arm_Request, Arm_Response,
+        StartTrajectory, StartTrajectory_Request, StartTrajectory_Response,
+        UploadTrajectory, UploadTrajectory_Request, UploadTrajectory_Response,
+        NotifySetpointsStop, NotifySetpointsStop_Request, NotifySetpointsStop_Response,
+    },
+    msg::{
+        Position,
+        VelocityWorld,
+        Hover,
+        FullState,
+        Status,
+    },
+};
+use std_srvs::srv::{Empty, Empty_Request, Empty_Response};
+use geometry_msgs::msg::{Twist, PoseStamped, Quaternion};
+use sensor_msgs::msg::LaserScan;
+use nav_msgs::msg::Odometry;
+use std_msgs::msg::String as StringMsg;
 use std::sync::Arc;
+
+mod config;
+use config::CrazyfliesConfig;
 
 
 struct CrazyflieROS {
-    node: Node,
+    _node: Node,
     cf: Arc<crazyflie_lib::Crazyflie>,
-    service_takeoff: Service<Takeoff>,
+    cf_name: String,
+    // Services
+    _service_emergency: Service<Empty>,
+    _service_arm: Service<Arm>,
+    _service_takeoff: Service<Takeoff>,
+    _service_land: Service<Land>,
+    _service_go_to: Service<GoTo>,
+    _service_start_trajectory: Service<StartTrajectory>,
+    _service_upload_trajectory: Service<UploadTrajectory>,
+    _service_notify_setpoints_stop: Service<NotifySetpointsStop>,
+    // Subscriptions
+    _sub_cmd_vel_legacy: Subscription<Twist>,
+    _sub_cmd_position: Subscription<Position>,
+    _sub_cmd_velocity_world: Subscription<VelocityWorld>,
+    _sub_cmd_hover: Subscription<Hover>,
+    _sub_cmd_full_state: Subscription<FullState>,
+    // Publishers
+    _pub_robot_description: Publisher<StringMsg>,
+    _pub_pose: Option<Publisher<PoseStamped>>,
+    _pub_scan: Option<Publisher<LaserScan>>,
+    _pub_odom: Option<Publisher<Odometry>>,
+    _pub_status: Option<Publisher<Status>>,
 }
 
 impl CrazyflieROS {
-    pub async fn new(node: Node, link_context: &crazyflie_link::LinkContext, uri: &str) -> Self {
+    pub async fn new(
+        node: Node, 
+        link_context: &crazyflie_link::LinkContext, 
+        uri: &str,
+        cf_name: String,
+    ) -> Result<Self, Box<dyn std::error::Error>> {
+        // Connect to Crazyflie
         let cf = crazyflie_lib::Crazyflie::connect_from_uri(
-                &link_context,
-                uri).await.unwrap();
-        println!("Connected!");
+            &link_context,
+            uri,
+        ).await?;
+        
+        println!("[{}] Connected!", cf_name);
         let cfarc = Arc::<Crazyflie>::new(cf);
 
-        let result = CrazyflieROS {
-            node: node.clone(),
-            cf: cfarc.clone(),
[…]
