+++ crazyflie_server_rs/Cargo.toml
@@ -5,7 +5,15 @@ edition = "2021"
 
 [dependencies]
 tokio = { version = "1", features = ["macros", "rt", "rt-multi-thread", "time"] }
-rclrs = "0.6"
+rclrs = "0.7"
 crazyflie-link = "0.3.0"
 crazyflie-lib = "0.3.0"
 crazyflie_interfaces = "*"
+std_srvs = "*"
+geometry_msgs = "*"
+sensor_msgs = "*"
+nav_msgs = "*"
+std_msgs = "*"
+builtin_interfaces = "*"
+serde = { version = "1.0", features = ["derive"] }
+serde_yaml = "0.9"
