### 2026-05-26 dc158bb
add crazyflieLinkCpp dependency
