### 2026-05-26 2a31b6e
rust server feature complete first try
