+++ crazyflie_server_rs/src/main.rs
@@ -28,6 +28,10 @@ use std::sync::Arc;
 mod config;
 use config::CrazyfliesConfig;
 
+mod params;
+mod log_helpers;
+use log_helpers::LogDataExt;
+
 
 struct CrazyflieROS {
     _node: Node,
@@ -105,12 +109,19 @@ impl CrazyflieROS {
         let service_arm = node.create_async_service::<Arm, _>(
             &format!("{}/arm", cf_name),
             move |request: Arm_Request| {
-                let _cf = cf_arm.clone();
+                let cf = cf_arm.clone();
                 async move {
                     println!("[Arm] arm={}", request.arm);
-                    // In the Python version, arming is done via parameter setting
-                    // For now, we'll just acknowledge the request
-                    // TODO: Implement parameter setting when available
+                    // Set the arm parameter via firmware param
+                    // Note: stabilizer.stop = 0 means armed, 1 means stopped
+                    let value = if request.arm {
+                        crazyflie_lib::Value::U8(0)  // Armed
+                    } else {
+                        crazyflie_lib::Value::U8(1)  // Stopped
+                    };
+                    if let Err(e) = cf.param.set("stabilizer.stop", value).await {
+                        eprintln!("Failed to set arm parameter: {}", e);
+                    }
                     Arm_Response::default()
                 }
             },
@@ -244,9 +255,17 @@ impl CrazyflieROS {
                         name,
                         request.trajectory_id, request.piece_offset, request.pieces.len()
                     );
-                    // TODO: Implement trajectory upload using memory subsystem
-                    // This requires converting TrajectoryPolynomialPiece to Poly4D format
-                    eprintln!("[{}] Trajectory upload not yet implemented", name);
+                    // Trajectory upload using memory subsystem
+                    // Note: This is a stub - full implementation requires:
+                    // 1. Getting trajectory memory handle from cf.memory
+                    // 2. Converting TrajectoryPolynomialPiece to Poly4D format
+                    // 3. Writing to trajectory memory slot
+                    // The crazyflie-lib API for this is:
+                    //   let traj_mem = cf.memory.trajectory_memories().await?;
+                    //   traj_mem.write(id, &poly4d_segments).await?;
+                    eprintln!("[{}] Trajectory upload: Converting {} pieces to Poly4D format", name, request.pieces.len());
+                    eprintln!("[{}] Trajectory upload not fully implemented - requires memory subsystem integration", name);
+                    // TODO: Implement full conversion and upload
                     UploadTrajectory_Response::default()
                 }
             },
@@ -348,13 +367,26 @@ impl CrazyflieROS {
         let cf_cmd_full = cfarc.clone();
         let sub_cmd_full_state = node.create_subscription(
             &format!("{}/cmd_full_state", cf_name),
-            move |_msg: FullState| {
-                let _cf = cf_cmd_full.clone();
+            move |msg: FullState| {
+                let cf = cf_cmd_full.clone();
                 tokio::spawn(async move {
-                    // TODO: Implement full state setpoint
-                    // The crazyflie-lib may not have a direct setpoint_full_state method yet
-                    // This would need to be sent as a generic CRTP packet
-                    eprintln!("cmd_full_state not yet fully implemented");
+                    // Full state setpoint: position + velocity + acceleration + attitude
+                    // crazyflie-lib doesn't have a direct full_state method, but we can use
[…]
