### 2026-05-27 615d06a
move simple mapper to examples
