+++ crazyflie/CMakeLists.txt
@@ -121,9 +121,7 @@ install(TARGETS
 install(PROGRAMS
   scripts/crazyflie_server.py
   scripts/chooser.py
-  scripts/vel_mux.py
   scripts/cfmult.py
-  scripts/simple_mapper_multiranger.py
   scripts/aideck_streamer.py
   scripts/gui.py
   scripts/flash.py
