+++ crazyflie/src/crazyflie_server.cpp
@@ -172,6 +172,12 @@ public:
     , tf_broadcaster_(node)
     , last_on_latency_(std::chrono::steady_clock::now())
     , cfbc_(cfbc)
+    , previous_numRxBc(0)
+    , previous_numRxUc(0)
+    , previous_stats_unicast_()
+    , previous_stats_broadcast_()
+    , last_latency_in_ms_(0)
+    , first_status_msg_(true)
   {
     auto sub_opt_cf_cmd = rclcpp::SubscriptionOptions();
     sub_opt_cf_cmd.callback_group = callback_group_cf_cmd;
@@ -197,6 +203,8 @@ public:
     subscription_cmd_position_ = node->create_subscription<crazyflie_interfaces::msg::Position>(name + "/cmd_position", rclcpp::SystemDefaultsQoS(), std::bind(&CrazyflieROS::cmd_position_changed, this, _1), sub_opt_cf_cmd);
     subscription_cmd_velocity_world_ = node->create_subscription<crazyflie_interfaces::msg::VelocityWorld>(name + "/cmd_velocity_world", rclcpp::SystemDefaultsQoS(), std::bind(&CrazyflieROS::cmd_velocity_world_changed, this, _1), sub_opt_cf_cmd);
     subscription_cmd_hover_ = node->create_subscription<crazyflie_interfaces::msg::Hover>(name + "/cmd_hover", rclcpp::SystemDefaultsQoS(), std::bind(&CrazyflieROS::cmd_hover_changed, this, _1), sub_opt_cf_cmd);
+    subscription_cmd_led_ = node->create_subscription<std_msgs::msg::String>(name + "/color_led",rclcpp::SystemDefaultsQoS(),std::bind(&CrazyflieROS::cmd_color_led_changed, this, _1),sub_opt_cf_cmd);
+
 
     publisher_robot_description_ = node->create_publisher<std_msgs::msg::String>(name + "/robot_description",
       rclcpp::QoS(1).transient_local());
@@ -548,7 +556,7 @@ public:
     }
 
     RCLCPP_INFO(logger_, "[%s] Requesting memories...", name_.c_str());
-    cf_.requestMemoryToc();
+    // cf_.requestMemoryToc();
   }
 
   void spin_once()
@@ -632,6 +640,8 @@ public:
     } else {
       RCLCPP_ERROR(logger_, "[%s] Could not find param %s/%s", name_.c_str(), group.c_str(), name.c_str());
     }
+
+
   }
 
 private:
@@ -685,6 +695,70 @@ private:
     float yawRate = -1.0 * msg->yaw_rate * 180.0 / M_PI; // Convert from radians to degrees
     float z = msg->z_distance;
     cf_.sendHoverSetpoint(vx, vy, yawRate, z);
+  }
+
+  void cmd_color_led_changed(const std_msgs::msg::String::SharedPtr msg)
+  {
+    std::string hex = msg->data;
+
+    if (!hex.empty() && hex[0] == '#') {
+      hex.erase(0, 1);
+    }
+
+    if (hex.rfind("0x", 0) == 0 || hex.rfind("0X", 0) == 0) {
+      hex.erase(0, 2);
+    }
+
+    try {
+        uint32_t raw_color = static_cast<uint32_t>(std::stoul(hex, nullptr, 16));
+
+        if (raw_color > 0xFFFFFF) {
+          RCLCPP_ERROR(logger_, "[%s] Invalid LED color \"%s\". Expected RRGGBB.", name_.c_str(), msg->data.c_str());
+          return;
+        }
+
+        // brightness entre 0.0 et 1.0
+        float brightness = 0.9f;   // 20% intensity
+
+        uint8_t r = (raw_color >> 16) & 0xFF;
+        uint8_t g = (raw_color >> 8)  & 0xFF;
[…]
