### 2026-05-27 4b4bf83
Reverting crazyflies.yaml back to the original
