+++ crazyflie/config/crazyflies.yaml
@@ -1,83 +1,12 @@
 # named list of all robots
+fileversion: 3
+
 robots:
-  C04:
-    enabled: false
-    uri: radio://0/80/1M/E7E7E7E704
-    initial_position: [0.0, 0.0, 0.0]
-    type: C10X # see robot_types
-  C05:
-    enabled: false
-    uri: radio://0/80/1M/E7E7E7E705
-    initial_position: [0.0, 0.0, 0.0]
-    type: C10X # see robot_types
-  C13:
-    enabled: false
-    uri: radio://0/80/1M/E7E7E7E70D
-    initial_position: [0.0, 0.0, 0.0]
-    type: C10X # see robot_types
-  C14:
-    enabled: false
-    uri: radio://0/80/1M/E7E7E7E70E
-    initial_position: [0.0, 0.0, 0.0]
-    type: C10X # see robot_types
-  C20:
-    enabled: false
-    uri: radio://0/80/2M/E7E7E7E714
-    initial_position: [0.0, 0.0, 0.0]
-    type: C10X # see robot_types
-  #crazyflies brushless #######################################################
-  C21:
-    enabled: false
-    uri: radio://0/80/2M/E7E7E7E715
-    initial_position: [0.0, 0.0, 0.0]
-    type: C10X # see robot_types
-    role: None
-  C22:
-    enabled: false
-    uri: radio://0/80/2M/E7E7E7E716
-    initial_position: [0.0, 0.0, 0.0]
-    type: C10X # see robot_types
-    role: None
-  C23:
+  cf231:
     enabled: true
-    uri: radio://0/80/2M/E7E7E7E717
+    uri: radio://0/80/2M/E7E7E7E7E7
     initial_position: [0.0, 0.0, 0.0]
-    type: C10X # see robot_types
-    role: evader
-  C24:
-    enabled: true
-    uri: radio://0/80/2M/E7E7E7E718
-    initial_position: [0.0, 0.0, 0.0]
-    type: C10X # see robot_types
-    role: pursuer
-  C25:
-    enabled: true
-    uri: radio://0/80/2M/E7E7E7E719
-    initial_position: [0.0, 0.0, 0.0]
-    type: C10X # s
-    role: pursuer
-  C26:
-    enabled: true
-    uri: radio://0/80/2M/E7E7E7E71A
-    initial_position: [0.0, 0.0, 0.0]
-    type: C10X # s
-    role: pursuer
-#####################################################################################
-  C101:
-    enabled: false
-    uri: radio://0/80/1M/E7E7E7E701
-    initial_position: [0.0, 0.0, 0.0]
-    type: C10X # see robot_types
-  C102: 
-    enabled: false
-    uri: radio://0/80/1M/E7E7E7E702
-    initial_position: [0.0, 0.0, 0.0]
-    type: C10X # see robot_types
-  C103:
-    enabled: false
-    uri: radio://0/80/1M/E7E7E7E703
-    initial_position: [0.0, -0.05, 0.0]
-    type: C10X  # see robot_types
+    type: cf21  # see robot_types
     # firmware_params:
     #   kalman:
     #     pNAcc_xy: 1.0 # default 0.5
@@ -87,80 +16,35 @@ robots:
     #     topic_name3: 
     #       frequency: 1
     #       vars: ["acc.x", "acc.y"]
-  C106:
+    # reference_frame: "odom_cf231" # use a custom reference frame if no global coordinate system is available
+
+  cf5:
     enabled: false
-    uri: radio://0/80/1M/E7E7E7E706
-    initial_position: [0.0, 0.0, 0.0]
[…]
