### 2026-06-02 2a076f6
remove from cmakelist and change launch file
