### 2026-06-05 4533dd2
add crazyflie_description to ci
