+++ .github/workflows/ci-ros2.yml
@@ -73,5 +73,6 @@ jobs:
             crazyflie_interfaces
             crazyflie_py
             crazyflie_sim
+            crazyflie_description
           target-ros2-distro: ${{ matrix.ros_distribution }}
           vcs-repo-file-url: rosinstall
