### 2026-06-09 e1f8e8c
feat/led deck commands and reboot service
