### 2026-06-10 8ee18a0
fix/fixing unitialized stats count that may cause incorrect statistics warnings
