### 2026-07-19 279b476
cpp-server: fix compiler error on rolling
